@@ -19,7 +19,7 @@
     "issues": "https://github.com/imbo/behat-api-extension/issues"
   },
   "require": {
-    "php": ">=8.1",
+    "php": ">=8.3",
     "ext-json": "*",
     "beberlei/assert": "^3.3",
     "behat/behat": "^3.8",
@@ -27,15 +27,16 @@
     "guzzlehttp/guzzle": "^7.3"
   },
   "require-dev": {
-    "alexeyshockov/guzzle-psalm-plugin": "^1.0",
     "imbo/imbo-coding-standard": "^2.0",
-    "phpunit/phpunit": "^10.0",
-    "psalm/plugin-phpunit": "^0.18.4",
+    "phpstan/extension-installer": "^1.4",
+    "phpstan/phpstan": "^2.1",
+    "phpstan/phpstan-deprecation-rules": "^2.0",
+    "phpstan/phpstan-phpunit": "^2.0",
+    "phpunit/phpunit": "^11.5",
     "slim/psr7": "^1.3",
     "slim/slim": "^4.7",
-    "symfony/process": "^6.2",
-    "tuupola/slim-basic-auth": "^3.3",
-    "vimeo/psalm": "^5.8"
+    "symfony/process": "^7.2",
+    "tuupola/slim-basic-auth": "^3.3"
   },
   "autoload": {
     "psr-4": {
@@ -61,12 +62,15 @@
     "phpunit": "vendor/bin/phpunit",
     "phpunit:coverage": "vendor/bin/phpunit --coverage-html build/coverage",
     "behat": "vendor/bin/behat --strict",
-    "sa": "vendor/bin/psalm",
+    "sa": "vendor/bin/phpstan",
     "cs": "php-cs-fixer fix --dry-run --diff",
     "dev": "php -S localhost:8080 -t ./features/bootstrap > server.log 2>&1",
     "docs": "cd docs; make html"
   },
   "config": {
-    "sort-packages": true
+    "sort-packages": true,
+    "allow-plugins": {
+      "phpstan/extension-installer": true
+    }
   }
 }
