@@ -2,11 +2,11 @@
 namespace Imbo\BehatApiExtension\ArrayContainsComparator\Matcher;
 
 use InvalidArgumentException;
+use PHPUnit\Framework\Attributes\CoversClass;
+use PHPUnit\Framework\Attributes\DataProvider;
 use PHPUnit\Framework\TestCase;
 
-/**
- * @coversDefaultClass Imbo\BehatApiExtension\ArrayContainsComparator\Matcher\ArrayMaxLength
- */
+#[CoversClass(ArrayMaxLength::class)]
 class ArrayMaxLengthTest extends TestCase
 {
     private ArrayMaxLength $matcher;
@@ -56,22 +56,16 @@ class ArrayMaxLengthTest extends TestCase
         ];
     }
 
-    /**
-     * @dataProvider getArraysAndLengths
-     * @covers ::__invoke
-     */
-    public function testCanMatchMaxLengthOfArrays(array $array, int $length): void
+    #[DataProvider('getArraysAndLengths')]
+    public function testCanMatchMaxLengthOfArrays(array $list, int $length): void
     {
         $matcher = $this->matcher;
         $this->assertTrue(
-            $matcher($array, $length),
+            $matcher($list, $length),
             'Matcher is supposed to return true.',
         );
     }
 
-    /**
-     * @covers ::__invoke
-     */
     public function testThrowsExceptionWhenMatchingAgainstAnythingOtherThanAnArray(): void
     {
         $this->expectException(InvalidArgumentException::class);
@@ -80,10 +74,7 @@ class ArrayMaxLengthTest extends TestCase
         $matcher(['foo' => 'bar'], 123);
     }
 
-    /**
-     * @dataProvider getValuesThatFail
-     * @covers ::__invoke
-     */
+    #[DataProvider('getValuesThatFail')]
     public function testThrowsExceptionWhenLengthIsTooShort(array $array, int $maxLength, string $message): void
     {
         $this->expectException(InvalidArgumentException::class);
